requirements.txt
amqp==2.1.4
appdirs==1.4.3
asn1crypto==0.22.0
Babel==2.4.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
click==6.7
coreapi==2.3.0
coreschema==0.0.4
cryptography==1.8.1
enum34==1.1.6
Fabric==1.13.1
Flask==0.12.1
flower==0.9.1
futures==3.1.1
honcho==1.0.1
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.9.6
kombu==4.0.2
MarkupSafe==1.0
packaging==16.8
paramiko==2.1.2
py==1.4.33
pyaml==16.12.2
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.2.0
pytest==3.0.7
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.13.0
six==1.10.0
tornado==4.2
ujson==1.35
uritemplate==3.0.0
vine==1.1.3
Werkzeug==0.12.1
